refactor: update folder structure with screens and correct imports
