+++ app/screens/profile.tsx
@@ -3,10 +3,10 @@ import React from "react";
 import { Text, View, SafeAreaView, Pressable, Alert, ScrollView } from "react-native";
 import { useRouter } from "expo-router";
 import { Feather } from "@expo/vector-icons";
-import { useTheme } from "./context/ThemeContext";
-import { useAuth } from "./context/AuthContext";
-import MacroSummary from "./components/MacroSummary";
-import { resetUserMacros } from "./models/user";
+import { useTheme } from "../context/ThemeContext";
+import { useAuth } from "../context/AuthContext";
+import MacroSummary from "../components/MacroSummary";
+import { resetUserMacros } from "../models/user";
 
 export default function Profile() {
   const router = useRouter();
